revert

+++ include/c-game/ctools.h
@@ -23,7 +23,7 @@
 
 /* get rid of these data type */
 typedef unsigned char uchar;
-typedef unsigned int uint;
+typedef unsigned short ushort;
 typedef unsigned int uint;
 
 /* macros */
